move testing suites to KafkaSourceProviderSuite
