+++ connector/kafka-0-10-sql/src/test/scala/org/apache/spark/sql/kafka010/KafkaRelationSuite.scala
@@ -19,21 +19,17 @@ package org.apache.spark.sql.kafka010
 
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.Locale
-import java.util.concurrent.ExecutionException
 import java.util.concurrent.atomic.AtomicInteger
 
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.TopicPartition
 
-import org.apache.spark.{SparkConf, SparkException, TestUtils}
-import org.apache.spark.sql.{DataFrameReader, QueryTest, Row}
+import org.apache.spark.{SparkConf, TestUtils}
+import org.apache.spark.sql.{DataFrameReader, QueryTest}
 import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.execution.datasources.v2.DataSourceV2Relation
-import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.streaming.StreamingQueryException
 import org.apache.spark.sql.test.SharedSparkSession
-import org.apache.spark.sql.types.{StringType, StructType}
 import org.apache.spark.util.Utils
 
 abstract class KafkaRelationSuiteBase extends QueryTest with SharedSparkSession with KafkaTest {
@@ -648,8 +644,6 @@ class KafkaRelationSuiteV1 extends KafkaRelationSuiteBase {
 }
 
 class KafkaRelationSuiteV2 extends KafkaRelationSuiteBase {
-  import scala.collection.JavaConverters._
-
   override protected def sparkConf: SparkConf =
     super
       .sparkConf
@@ -661,60 +655,5 @@ class KafkaRelationSuiteV2 extends KafkaRelationSuiteBase {
     assert(df.logicalPlan.collect {
       case _: DataSourceV2Relation => true
     }.nonEmpty)
-  }
-
-  Seq("source and stream", "sink and stream",
-    "source and batch", "sink and batch").foreach { testType =>
-    val options: Map[String, String] = Map(
-      "kafka.bootstrap.servers" -> "localhost:8080",
-      "subscribe" -> "msk-123",
-      "startingOffsets" -> "earliest",
-      "kafka.sasl.mechanism" -> "AWS_MSK_IAM",
-      "kafka.sasl.jaas.config" ->
-        "software.amazon.msk.auth.iam.IAMLoginModule required;",
-      "kafka.security.protocol" -> "SASL_SSL",
-      "kafka.sasl.client.callback.handler.class" ->
-        "software.amazon.msk.auth.iam.IAMClientCallbackHandler"
-    )
-
-    // if testType contains source/sink, then kafka is used as a source/sink option respectively
-    // if testType contains stream/batch, it is used either in readStream/read
-    test(s"test MSK IAM auth on kafka '$testType' side") {
-      var e : Throwable = null
-      if (testType.contains("stream")) {
-          if (testType.contains("source")) {
-            e = intercept[StreamingQueryException] {
-              spark.readStream.format("kafka").options(options).load()
-                .writeStream.format("console").start().processAllAvailable()
-            }
-            TestUtils.assertExceptionMsg(e, "Timed out")
-          } else {
-            e = intercept[StreamingQueryException] {
-              spark.readStream.format("rate").option("rowsPerSecond", 10).load()
-                .withColumn("value", col("value").cast(StringType)).writeStream
-                .format("kafka").options(options).option("checkpointLocation", "temp/testing")
-                .option("topic", "msk-123").start().processAllAvailable()
-            }
-            TestUtils.assertExceptionMsg(e, "Timeout")
-          }
-      } else {
-        if (testType.contains("source")) {
-          e = intercept[ExecutionException] {
-            spark.read.format("kafka").options(options).load()
-              .write.format("console").save()
-          }
-          TestUtils.assertExceptionMsg(e, "Timed out")
-        } else {
-          val schema = new StructType().add("value", "string")
-          e = intercept[SparkException] {
-            spark.createDataFrame(Seq(Row("test"), Row("test2")).asJava, schema)
-              .write.mode("append").format("kafka")
-              .options(options).option("checkpointLocation", "temp/testing/1")
-              .option("topic", "msk-123").save()
-          }
-          TestUtils.assertExceptionMsg(e, "Timeout")
-        }
-      }
-    }
   }
 }
